+++ lib/src/model/pluto_column.dart
@@ -145,8 +145,12 @@ class PlutoColumn {
 
   /// A checkbox appears in the cell of the column.
   bool enableRowChecked;
-  int rowCheckBoxGroupDepth; //
+
+  int rowCheckBoxGroupDepth;
+
   bool enableTitleChecked;
+
+  bool Function(PlutoRow row)? disableRowCheckboxWhen;
 
   /// Sort rows by tapping on the column heading.
   bool enableSorting;
@@ -176,27 +180,6 @@ class PlutoColumn {
   /// Valid only when [enableContextMenu] is activated.
   bool enableFilterMenuItem;
 
-  ///Set hint text for filter field
-  String? filterHintText;
-
-  ///Set hint text color for filter field
-  Color? filterHintTextColor;
-
-  ///Set suffix icon for filter field
-  Icon? filterSuffixIcon;
-
-  ///Set custom widget
-  @Deprecated("Use new filterWidgetBuilder to provide some parameters")
-  Widget? filterWidget;
-
-  Widget Function(
-    FocusNode focusNode,
-    TextEditingController controller,
-    bool enabled,
-    void Function(String changed) handleOnChanged,
-    PlutoGridStateManager stateManager,
-  )? filterWidgetBuilder;
-
   /// Displays Hide column menu in the column context menu.
   /// Valid only when [enableContextMenu] is activated.
   bool enableHideColumnMenuItem;
@@ -215,6 +198,9 @@ class PlutoColumn {
 
   PlutoOnRowDoubleTapEventCallback? onCellDoubleTap;
   LinearGradient? backgroundGradient;
+
+  /// The widget of the filter column, this can be customized with the multiple constructors, defaults to a [PlutoFilterColumnWidgetDelegate.initial()]
+  PlutoFilterColumnWidgetDelegate? filterWidgetDelegate;
 
   PlutoColumn({
     required this.title,
@@ -247,11 +233,6 @@ class PlutoColumn {
     this.enableContextMenu = true,
     this.enableDropToResize = true,
     this.enableFilterMenuItem = true,
-    this.filterHintText,
-    this.filterHintTextColor,
-    this.filterSuffixIcon,
-    @Deprecated("Use new filterWidgetBuilder to provide some parameters")
-    this.filterWidget,
     this.enableHideColumnMenuItem = true,
     this.enableSetColumnsMenuItem = true,
     this.enableAutoEditing = false,
@@ -259,7 +240,9 @@ class PlutoColumn {
     this.hide = false,
     this.onCellDoubleTap,
     this.backgroundGradient,
-    this.filterWidgetBuilder,
+    this.filterWidgetDelegate =
+        const PlutoFilterColumnWidgetDelegate.textField(),
+    this.disableRowCheckboxWhen,
   })  : _key = UniqueKey(),
         _checkReadOnly = checkReadOnly {
     this.enableRowDrag = enableRowDrag ?? (_, __) => false;
@@ -363,6 +346,60 @@ class PlutoColumn {
 
     return value.toString();
   }
+}
+
+class PlutoFilterColumnWidgetDelegate {
+  /// This is the default filter widget delegate
+  const PlutoFilterColumnWidgetDelegate.textField({
+    this.filterHintText,
+    this.filterHintTextColor,
+    this.filterSuffixIcon,
+    this.onFilterSuffixTap,
+    this.clearIcon = const Icon(Icons.clear),
+    this.onClear,
+  }) : filterWidgetBuilder = null;
+
+  /// If you don't want a custom widget
+  const PlutoFilterColumnWidgetDelegate.builder({
+    this.filterWidgetBuilder,
+  })  : filterSuffixIcon = null,
+        onFilterSuffixTap = null,
+        filterHintText = null,
+        filterHintTextColor = null,
+        clearIcon = const Icon(Icons.clear),
+        onClear = null;
+
+  ///Set hint text for filter field
+  final String? filterHintText;
+
+  ///Set hint text color for filter field
+  final Color? filterHintTextColor;
+
+  ///Set suffix icon for filter field
+  final Widget? filterSuffixIcon;
+
+  /// Clear icon in the text field, if onClear is null, this will not appear
+  final Widget clearIcon;
+
+  /// If this is set, it will be called when the clear button is tapped, if this is null there won't be a clear icon
+  final Function? onClear;
+
+  /// Set a custom on tap event for the filter suffix icon
+  final Function(
+    FocusNode focusNode,
+    TextEditingController controller,
+    bool enabled,
+    void Function(String changed) handleOnChanged,
+    PlutoGridStateManager stateManager,
+  )? onFilterSuffixTap;
+
+  final Widget Function(
+    FocusNode focusNode,
+    TextEditingController controller,
+    bool enabled,
+    void Function(String changed) handleOnChanged,
+    PlutoGridStateManager stateManager,
+  )? filterWidgetBuilder;
 }
 
 class PlutoColumnRendererContext {