+++ lib/src/pluto_grid_configuration.dart
@@ -232,8 +232,10 @@ class PlutoGridStyleConfig {
     this.cellColorInEditState = Colors.white,
     this.cellColorInReadOnlyState = const Color(0xFFDBDBDC),
     this.cellColorGroupedRow,
+    this.frozenRowColor = const Color(0xFFF8F8F8),
+    this.frozenRowBorderColor = const Color(0xFFE0E0E0),
     this.dragTargetColumnColor = const Color(0xFFDCF5FF),
-    this.iconColor = Colors.black26,
+    this.iconColor = Colors.black38,
     this.disabledIconColor = Colors.black12,
     this.menuBackgroundColor = Colors.white,
     this.gridBorderColor = const Color(0xFFA1A5AE),
@@ -307,6 +309,8 @@ class PlutoGridStyleConfig {
     this.cellColorInEditState = const Color(0xFF666666),
     this.cellColorInReadOnlyState = const Color(0xFF222222),
     this.cellColorGroupedRow,
+    this.frozenRowColor = const Color(0xFF222222),
+    this.frozenRowBorderColor = const Color(0xFF666666),
     this.dragTargetColumnColor = const Color(0xFF313131),
     this.iconColor = Colors.white38,
     this.disabledIconColor = Colors.white12,
@@ -432,6 +436,12 @@ class PlutoGridStyleConfig {
 
   /// Background color of cells grouped by column.
   final Color? cellColorGroupedRow;
+
+  /// Background color for frozen rows
+  final Color frozenRowColor;
+
+  /// Border color for frozen rows
+  final Color frozenRowBorderColor;
 
   /// The background color of the column to be dragged.
   /// When moving a column by dragging it.
@@ -636,7 +646,8 @@ class PlutoGridStyleConfig {
         defaultColumnFilterPadding: defaultColumnFilterPadding ?? this.defaultColumnFilterPadding,
         defaultCellPadding: defaultCellPadding ?? this.defaultCellPadding,
         columnTextStyle: columnTextStyle ?? this.columnTextStyle,
-        columnUnselectedColor: columnUnselectedColor ?? this.columnUnselectedColor,
+        columnUnselectedColor:
+            columnUnselectedColor ?? this.columnUnselectedColor,
         columnActiveColor: columnActiveColor ?? this.columnActiveColor,
         cellUnselectedColor: cellUnselectedColor ?? this.cellUnselectedColor,
         cellActiveColor: cellActiveColor ?? this.cellActiveColor,
@@ -1765,12 +1776,16 @@ class PlutoGridLocaleText {
 enum PlutoGridRowSelectionCheckBoxBehavior {
   /// Selecting a row does nothing to its checkbox
   none,
+
   /// Automatically enables the checkbox of the selected rows
   checkRow,
+
   /// Automatically toggles the checkbox of the selected rows
   toggleCheckRow,
+
   /// Automatically enabels the checkbox of a selected row (if another row is checked via select, the previous one is unchecked)
   singleRowCheck,
+
   /// Automatically toggles the checkbox of a selected row (if another row is checked via select, the previous one is unchecked)
   toggleSingleRowCheck,
 }